+++ .gitignore
@@ -350,3 +350,4 @@ fpga/.Xil/
 # vscode temp file
 .vscode/
 .bloop
+.metals/
