+++ src/main/scala/system/NutShell.scala
@@ -17,6 +17,8 @@
 package system
 
 import nutcore._
+import nutcore.mem.cache._
+
 import bus.axi4.{AXI4, AXI4Lite}
 import bus.simplebus._
 import device.{AXI4CLINT, AXI4PLIC}
