+++ src/main/scala/system/Prefetcher.scala
@@ -16,7 +16,9 @@
 
 package system
 
-import nutcore.{NutCore, NutCoreConfig, HasNutCoreParameter, AddressSpace, Cache, CacheConfig}
+import nutcore.{NutCore, NutCoreConfig, HasNutCoreParameter, AddressSpace}
+import nutcore.mem.cache._
+
 import bus.axi4.{AXI4, AXI4Lite}
 import bus.simplebus._
 import utils._
